# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/tournaments": {
      "get": {
        "operationId": "TournamentsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Public Tournament page; every item includes derived favoriteCount and viewer-specific isFavorited.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/TournamentFavoriteViewFieldsDto"
                      }
                    }
                  },
                  "additionalProperties": true
                }
              }
            }
          }
        },
        "tags": [
          "Tournaments"
        ]
      },
      "post": {
        "operationId": "TournamentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTournamentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Tournaments"
        ]
      }
    },
    "/api/tournaments/slug/{slug}": {
      "get": {
        "operationId": "TournamentsController_findByLegacySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tournament detail including derived favoriteCount and viewer-specific isFavorited.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TournamentFavoriteViewFieldsDto"
                }
              }
            }
          }
        },
        "tags": [
          "Tournaments"
        ]
      }
    },
    "/api/tournaments/{slug}/favorite": {
      "post": {
        "operationId": "TournamentsController_favorite",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TournamentFavoriteMutationResultDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Favorite and follow important Tournament-wide updates",
        "tags": [
          "Tournaments"
        ]
      },
      "delete": {
        "operationId": "TournamentsController_unfavorite",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TournamentFavoriteMutationResultDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove a Favorite and stop future follower notifications",
        "tags": [
          "Tournaments"
        ]
      }
    },
    "/api/tournaments/{slug}/standings": {
      "get": {
        "operationId": "TournamentsController_standings",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tournaments"
        ]
      }
    },
    "/api/tournaments/{slug}/schedule": {
      "get": {
        "operationId": "TournamentsController_schedule",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tournaments"
        ]
      }
    },
    "/api/tournaments/{slug}/bracket": {
      "get": {
        "operationId": "TournamentsController_bracket",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tournaments"
        ]
      }
    },
    "/api/tournaments/{slug}": {
      "get": {
        "operationId": "TournamentsController_findBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tournament detail including derived favoriteCount and viewer-specific isFavorited.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TournamentFavoriteViewFieldsDto"
                }
              }
            }
          }
        },
        "tags": [
          "Tournaments"
        ]
      }
    },
    "/api/tournaments/{tournamentId}": {
      "patch": {
        "operationId": "TournamentsController_update",
        "parameters": [
          {
            "name": "tournamentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTournamentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tournaments"
        ]
      },
      "delete": {
        "operationId": "TournamentsController_remove",
        "parameters": [
          {
            "name": "tournamentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tournaments"
        ]
      }
    },
    "/api/tournaments/{slug}/rounds": {
      "post": {
        "operationId": "TournamentsController_addRound",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoundDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Tournaments"
        ]
      }
    },
...
    "/api/users/me/avatar": {
      "post": {
        "operationId": "UploadController_avatar",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "uploads"
        ]
      }
    },
    "/api/teams/{id}/logo": {
      "post": {
        "operationId": "UploadController_teamLogo",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "uploads"
        ]
      }
    },
    "/api/teams/{id}/members/{memberId}/avatar": {
      "post": {
        "operationId": "UploadController_memberAvatar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "memberId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "uploads"
        ]
      }
    },
    "/api/tournaments/{tournamentId}/banner": {
      "post": {
        "operationId": "UploadController_banner",
        "parameters": [
          {
            "name": "tournamentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "uploads"
        ]
      }
    }
  },
  "info": {
    "title": "Esports Tournament Manager API",
    "description": "Backend API contract",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "TournamentFavoriteViewFieldsDto": {
        "type": "object",
        "properties": {
          "favoriteCount": {
            "type": "number",
            "example": 42,
            "minimum": 0
          },
          "isFavorited": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "favoriteCount",
          "isFavorited"
        ]
      },
...
      "TournamentFavoriteMutationResultDto": {
        "type": "object",
        "properties": {
          "isFavorited": {
            "type": "boolean",
            "example": true
          },
          "favoriteCount": {
            "type": "number",
            "example": 42,
            "minimum": 0
          }
        },
        "required": [
          "isFavorited",
          "favoriteCount"
        ]
      },
      "CreateTournamentDto": {
        "type": "object",
        "properties": {
          "teamSize": {
            "type": "number",
            "description": "Số người thi đấu chính mỗi bên; backend kiểm tra theo metadata của Game",
            "minimum": 1,
            "maximum": 50
          },
          "customGameName": {
            "type": "string",
            "description": "Tên game do organizer đặt khi chọn canonical game CUSTOM",
            "maxLength": 100
          }
        }
      },
      "UpdateTournamentDto": {
        "type": "object",
        "properties": {
          "teamSize": {
            "type": "number",
            "description": "Số người thi đấu chính mỗi bên; backend kiểm tra theo metadata của Game",
            "minimum": 1,
            "maximum": 50
          },
          "customGameName": {
            "type": "string",
            "description": "Tên game do organizer đặt khi chọn canonical game CUSTOM",
            "maxLength": 100
          }
        }
      },
      "CreateRoundDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}